fixed validation errors and migration inputs

+++ server/db/migrations/groupBenefit.js
@@ -1,9 +1,9 @@
 exports.up = (knex, Promise) => {
   return knex.schema.createTable('group_benefit', (table) => {
     table.increments('id').primary()
-    table.integer('pessimistic')
-    table.integer('optimistic')
-    table.integer('likely')
+    table.decimal('pessimistic')
+    table.decimal('optimistic')
+    table.decimal('likely')
     table.integer('chance_of_success')
     table.integer('group_id')
   })
